+++ src/citra/citra.cpp
@@ -76,8 +76,7 @@ static void PrintHelp(const char* argv0) {
 }
 
 static void PrintVersion() {
-    std::cout << "Citra Valentin " << Version::major << "." << Version::minor << "."
-              << Version::patch << std::endl;
+    std::cout << "Citra Valentin " << Version::citra_valentin.to_string() << std::endl;
 }
 
 static void OnStateChanged(const Network::RoomMember::State& state) {