+++ src/citra_qt/configuration/configure_versions_linux.cpp
@@ -1,0 +1,10 @@
+// Copyright 2019 Citra Valentin Emulator Project
+// Licensed under GPLv2 or any later version
+// Refer to the license.txt file included.
+
+#include "citra_qt/configuration/configure_versions.h"
+#include "ui_configure_versions.h"
+
+void ConfigureVersions::Initialize() {}
+
+void ConfigureVersions::UpdateInstalledVersions() {}